--- =7.2.0
@@ -1,8 +1,0 @@
-+ @babel/core@7.11.6
-added 4 packages from 1 contributor, removed 139 packages, updated 3 packages and audited 213 packages in 2.429s
-
-27 packages are looking for funding
-  run `npm fund` for details
-
-found 0 vulnerabilities
-
