+++ distances/distance_metric.js
